+++ NEWS.md
@@ -1,3 +1,10 @@
+# readit version 1.1.0
+
+This version adds deprecation warnings via `.Deprecated()` to all calls to
+`readit()`, pointing users instead to the `rio` package, due to its much better
+design and enhanced functionality. The `readit` package is retired, and will not
+be developed further.
+
 # readit version 1.0.0
 
 This is the first CRAN release of the readit package. Included is the readit()